requirements.txt
torch
torchvision
opencv-python
numpy
pandas
matplotlib
scikit-learn
tqdm
Pillow
Flask
Flask-CORS
werkzeug
cython
pycocotools
pyyaml
argparse
json5
scipy
pymongo
flask-login
flask-bcrypt
python-dotenv
jwt
PyJWT
requests

# Detectron2 needs to be installed separately with:
# pip install detectron2 -f https://dl.fbaipublicfiles.com/detectron2/wheels/cu111/torch1.8/index.html
# Or for the latest version:
# pip install --no-isolation 'git+https://github.com/facebookresearch/detectron2.git'